Refactor `tasks` to `subscribers` because of celery auto-loading - fixes #23
